A0-4224: Extracted pallet alleph runtime API to a separate crate
