+++ bin/node/Cargo.toml
@@ -66,6 +66,8 @@ aleph-runtime = { workspace = true }
 aleph-runtime-interfaces = { workspace = true }
 finality-aleph = { workspace = true }
 primitives = { workspace = true }
+fake-runtime-api = { workspace = true, features = ["std"] }
+pallet-aleph-runtime-api = { workspace = true, features = ["std"] }
 
 # These dependencies are used for the node's RPCs
 jsonrpsee = { workspace = true, features = ["server"] }
