+++ bin/runtime/Cargo.toml
@@ -61,6 +61,7 @@ sp-genesis-builder = { workspace = true }
 frame-benchmarking = { workspace = true, optional = true }
 
 pallet-aleph = { workspace = true }
+pallet-aleph-runtime-api = { workspace = true }
 pallet-committee-management = { workspace = true }
 pallet-elections = { workspace = true }
 pallet-feature-control = { workspace = true }
