+++ finality-aleph/src/party/manager/mod.rs
@@ -7,13 +7,14 @@ use network_clique::SpawnHandleT;
 use sc_keystore::{Keystore, LocalKeystore};
 use sp_application_crypto::RuntimeAppPublic;
 use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
+use pallet_aleph_runtime_api::AlephSessionApi;
 
 use crate::{
     abft::{
         current_create_aleph_config, legacy_create_aleph_config, run_current_member,
         run_legacy_member, SpawnHandle,
     },
-    aleph_primitives::{AlephSessionApi, BlockHash, BlockNumber, KEY_TYPE},
+    aleph_primitives::{BlockHash, BlockNumber, KEY_TYPE},
     block::{
         substrate::{Justification, JustificationTranslator},
         BestBlockSelector, Block, Header, HeaderVerifier, UnverifiedHeader,
