+++ pallets/aleph-runtime-api/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "pallet-aleph-runtime-api"
+version = "0.1.0"
+license = "Apache 2.0"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+sp-api = { workspace = true }
+sp-std = { workspace = true }
+sp-consensus-aura = { workspace = true }
+
+primitives = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+    "sp-api/std",
+    "sp-std/std",
+    "primitives/std",
+]
