+++ pallets/aleph-runtime-api/README.md
@@ -1,0 +1,5 @@
+# pallet-aleph-runtime-api
+
+Runtime API definition for aleph pallet.
+
+License: Apache-2.0
